APIs

DynamicQuantumCircuits.unitary_reconstruction — Function

unitary_reconstruction(qc_dynamic::MainProgram)::MainProgram

Reconstructs a unitary quantum circuit from a dynamic quantum circuit

DynamicQuantumCircuits.validate_state — Function

validate_state(upper::Vector{Any}, lower::Vector{Any})

The combined StateVector needs to be a Set

TODO Check if the the states are increase by one